refactor: input_util -> general
