+++ toolchest_client/files/__init__.py
@@ -1,5 +1,5 @@
-from .general import assert_exists, check_file_size, files_in_path, sanity_check, compress_files_in_path
-from .input_util import convert_input_params_to_prefix_mapping
+from .general import assert_exists, check_file_size, files_in_path, sanity_check, compress_files_in_path, \
+    convert_input_params_to_prefix_mapping
 from .merge import concatenate_files, merge_sam_files
 from .s3 import assert_accessible_s3, get_s3_file_size, get_params_from_s3_uri, path_is_s3_uri
 from .split import open_new_output_file, split_file_by_lines, split_paired_files_by_lines
