+++ toolchest_client/files/tests/test_general.py
@@ -3,7 +3,7 @@ import pathlib
 
 import pytest
 
-from .. import assert_exists, check_file_size, files_in_path, sanity_check
+from .. import assert_exists, check_file_size, files_in_path, sanity_check, convert_input_params_to_prefix_mapping
 
 THIS_FILE_PATH = os.path.normpath(pathlib.Path(__file__).parent.resolve())
 
@@ -48,3 +48,46 @@ def test_exists_but_not_file():
     dir_file_path = f"{THIS_FILE_PATH}/data"
     with pytest.raises(ValueError):
         assert_exists(dir_file_path, must_be_file=True)
+
+
+def test_generate_prefix_mapping():
+    tag_to_param_map = {
+        "-1": ["example1_R1.fastq", "example2_R1.fastq"],
+        "-2": ["example1_R2.fastq", "example2_R2.fastq"],
+        "-U": ["example1_U.fastq", "example2_U.fastq"],
+    }
+    input_list, prefix_mapping = convert_input_params_to_prefix_mapping(tag_to_param_map)
+    assert sorted(input_list) == sorted([
+        "example1_R1.fastq",
+        "example2_R1.fastq",
+        "example1_R2.fastq",
+        "example2_R2.fastq",
+        "example1_U.fastq",
+        "example2_U.fastq",
+    ])
+    assert prefix_mapping == {
+        "example1_R1.fastq": {
+            "prefix": "-1",
+            "order": 0,
+        },
+        "example1_R2.fastq": {
+            "prefix": "-2",
+            "order": 0,
+        },
+        "example1_U.fastq": {
+            "prefix": "-U",
+            "order": 0,
+        },
+        "example2_R1.fastq": {
+            "prefix": "-1",
+            "order": 1,
+        },
+        "example2_R2.fastq": {
+            "prefix": "-2",
+            "order": 1,
+        },
+        "example2_U.fastq": {
+            "prefix": "-U",
+            "order": 1,
+        },
+    }
