--- toolchest_client/files/tests/test_input_params.py
@@ -1,44 +1,0 @@
-from .. import convert_input_params_to_prefix_mapping
-
-
-def test_generate_prefix_mapping():
-    tag_to_param_map = {
-        "-1": ["example1_R1.fastq", "example2_R1.fastq"],
-        "-2": ["example1_R2.fastq", "example2_R2.fastq"],
-        "-U": ["example1_U.fastq", "example2_U.fastq"],
-    }
-    input_list, prefix_mapping = convert_input_params_to_prefix_mapping(tag_to_param_map)
-    assert sorted(input_list) == sorted([
-        "example1_R1.fastq",
-        "example2_R1.fastq",
-        "example1_R2.fastq",
-        "example2_R2.fastq",
-        "example1_U.fastq",
-        "example2_U.fastq",
-    ])
-    assert prefix_mapping == {
-        "example1_R1.fastq": {
-            "prefix": "-1",
-            "order": 0,
-        },
-        "example1_R2.fastq": {
-            "prefix": "-2",
-            "order": 0,
-        },
-        "example1_U.fastq": {
-            "prefix": "-U",
-            "order": 0,
-        },
-        "example2_R1.fastq": {
-            "prefix": "-1",
-            "order": 1,
-        },
-        "example2_R2.fastq": {
-            "prefix": "-2",
-            "order": 1,
-        },
-        "example2_U.fastq": {
-            "prefix": "-U",
-            "order": 1,
-        },
-    }
